refactor(library/Byte): extracts "cofactorTo.ts" from "index.ts"
